requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
Pillow>=8.3.1
tqdm>=4.62.3
matplotlib>=3.4.3
pycocotools>=2.0.2 